*** empty log message ***

+++ update-version.sh
@@ -12,18 +12,29 @@ fi
 
 VERSION=${1}
 
-cp configure.ac configure.ac.oldvers
-export TAG_CONF=AC_INIT
-awk -F ',' \
+if [ -f "configure.ac" ]; then
+  cp configure.ac configure.ac.oldvers
+  export TAG_CONF=AC_INIT
+  awk -F ',' \
     '$0  ~ /^'${TAG_CONF}'/ { print $1 ", '${VERSION}'," $3; } ; \
      $0 !~ /^'${TAG_CONF}'/ { print $0; }' configure.ac > configure.ac.newchgvers
-mv configure.ac.newchgvers configure.ac
+  mv configure.ac.newchgvers configure.ac
+  echo; echo "[UPDATE-VERSION] Changes in configure.ac:"
+  diff -s configure.ac.oldvers configure.ac
+else
+  echo; echo "[UPDATE-VERSION] ERROR: configure.ac not found!"
+fi
 
-
-cp paravermain.h paravermain.h.oldvers
-export TAG_CONF=VERSION
-awk -F ' ' \
+if [ -f "configure.ac" ]; then
+  cp paravermain.h paravermain.h.oldvers
+  export TAG_CONF=VERSION
+  awk -F ' ' \
      '$0  ~ /define '${TAG_CONF}'/ { print $1 " '${TAG_CONF}' \"'${VERSION}'\"" ; } ;  \
       $0 !~ /define '${TAG_CONF}'/ { print $0; }'  paravermain.h > paravermain.h.newchgvers
-mv paravermain.h.newchgvers paravermain.h
+  mv paravermain.h.newchgvers paravermain.h
+  echo; echo "[UPDATE-VERSION] Changes in paravermain.h:"
+  diff -s paravermain.h.oldvers paravermain.h
+else
+  echo; echo "[UPDATE-VERSION] ERROR: paravermain.h not found!"
+fi
